+++ bricks/ui_kit/__brick__/lib/constants/{{class_name.snakeCase()}}_constants.dart
@@ -1,0 +1,5 @@
+final class {{class_name.pascalCase()}}Constants {
+  {{class_name.pascalCase()}}Constants._();
+
+  static const String packageName = '{{class_name.snakeCase()}}';
+}