+++ bricks/ui_kit/__brick__/lib/layout/gap/{{class_name.snakeCase()}}_v_gap.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class {{class_name.pascalCase()}}VGap extends SizedBox {
+  /// Vertical Gap Zero
+  const {{class_name.pascalCase()}}VGap({super.key}) : super(height: 0);
+  const {{class_name.pascalCase()}}VGap.zero({super.key}) : super(height: 0);
+
+  /// Vertical Gap 4
+  const {{class_name.pascalCase()}}VGap.v4({super.key}) : super(height: 4);
+
+  /// Vertical Gap 8
+  const {{class_name.pascalCase()}}VGap.v8({super.key}) : super(height: 8);
+
+  /// Vertical Gap 12
+  const {{class_name.pascalCase()}}VGap.v12({super.key}) : super(height: 12);
+
+  /// Vertical Gap 16
+  const {{class_name.pascalCase()}}VGap.v16({super.key}) : super(height: 16);
+
+  /// Vertical Gap 20
+  const {{class_name.pascalCase()}}VGap.v20({super.key}) : super(height: 20);
+
+  /// Vertical Gap 24
+  const {{class_name.pascalCase()}}VGap.v24({super.key}) : super(height: 24);
+
+  /// Vertical Gap 28
+  const {{class_name.pascalCase()}}VGap.v28({super.key}) : super(height: 28);
+
+  /// Vertical Gap 32
+  const {{class_name.pascalCase()}}VGap.v32({super.key}) : super(height: 32);
+
+  /// Vertical Gap 36
+  const {{class_name.pascalCase()}}VGap.v36({super.key}) : super(height: 36);
+
+  /// Vertical Gap 40
+  const {{class_name.pascalCase()}}VGap.v40({super.key}) : super(height: 40);
+}