+++ bricks/ui_kit/__brick__/lib/layout/padding/{{class_name.snakeCase()}}_padding.dart
@@ -1,0 +1,224 @@
+import 'package:flutter/material.dart';
+
+class {{class_name.pascalCase()}}Padding extends EdgeInsetsDirectional {
+  /// Zero
+  const {{class_name.pascalCase()}}Padding() : super.all(0);
+  const {{class_name.pascalCase()}}Padding.zero() : super.all(0);
+
+  /// Padding all
+  /// Padding all 4
+  const {{class_name.pascalCase()}}Padding.all4() : super.all(4);
+
+  /// Padding all 8
+  const {{class_name.pascalCase()}}Padding.all8() : super.all(8);
+
+  /// Padding all 12
+  const {{class_name.pascalCase()}}Padding.all12() : super.all(12);
+
+  /// Padding all 16
+  const {{class_name.pascalCase()}}Padding.all16() : super.all(16);
+
+  /// Padding all 20
+  const {{class_name.pascalCase()}}Padding.all20() : super.all(20);
+
+  /// Padding all 24
+  const {{class_name.pascalCase()}}Padding.all24() : super.all(24);
+
+  /// Padding all 28
+  const {{class_name.pascalCase()}}Padding.all28() : super.all(28);
+
+  /// Padding all 32
+  const {{class_name.pascalCase()}}Padding.all32() : super.all(32);
+
+  /// Padding all 36
+  const {{class_name.pascalCase()}}Padding.all36() : super.all(36);
+
+  /// Padding all 40
+  const {{class_name.pascalCase()}}Padding.all40() : super.all(40);
+
+  /// Padding Horizonal
+  /// Padding Horizonal 4
+  const {{class_name.pascalCase()}}Padding.horizontal4() : super.symmetric(horizontal: 4);
+
+  /// Padding Horizonal 8
+  const {{class_name.pascalCase()}}Padding.horizontal8() : super.symmetric(horizontal: 8);
+
+  /// Padding Horizonal 12
+  const {{class_name.pascalCase()}}Padding.horizontal12() : super.symmetric(horizontal: 12);
+
+  /// Padding Horizonal 16
+  const {{class_name.pascalCase()}}Padding.horizontal16() : super.symmetric(horizontal: 16);
+
+  /// Padding Horizonal 20
+  const {{class_name.pascalCase()}}Padding.horizontal20() : super.symmetric(horizontal: 20);
+
+  /// Padding Horizonal 24
+  const {{class_name.pascalCase()}}Padding.horizontal24() : super.symmetric(horizontal: 24);
+
+  /// Padding Horizonal 28
+  const {{class_name.pascalCase()}}Padding.horizontal28() : super.symmetric(horizontal: 28);
+
+  /// Padding Horizonal 32
+  const {{class_name.pascalCase()}}Padding.horizontal32() : super.symmetric(horizontal: 32);
+
+  /// Padding Horizonal 36
+  const {{class_name.pascalCase()}}Padding.horizontal36() : super.symmetric(horizontal: 36);
+
+  /// Padding Horizonal 40
+  const {{class_name.pascalCase()}}Padding.horizontal40() : super.symmetric(horizontal: 40);
+
+  /// Padding Vertical
+  /// Padding Vertical 4
+  const {{class_name.pascalCase()}}Padding.vertical4() : super.symmetric(vertical: 4);
+
+  /// Padding Vertical 8
+  const {{class_name.pascalCase()}}Padding.vertical8() : super.symmetric(vertical: 8);
+
+  /// Padding Vertical 12
+  const {{class_name.pascalCase()}}Padding.vertical12() : super.symmetric(vertical: 12);
+
+  /// Padding Vertical 16
+  const {{class_name.pascalCase()}}Padding.vertical16() : super.symmetric(vertical: 16);
+
+  /// Padding Vertical 20
+  const {{class_name.pascalCase()}}Padding.vertical20() : super.symmetric(vertical: 20);
+
+  /// Padding Vertical 24
+  const {{class_name.pascalCase()}}Padding.vertical24() : super.symmetric(vertical: 24);
+
+  /// Padding Vertical 28
+  const {{class_name.pascalCase()}}Padding.vertical28() : super.symmetric(vertical: 28);
+
+  /// Padding Vertical 32
+  const {{class_name.pascalCase()}}Padding.vertical32() : super.symmetric(vertical: 32);
+
+  /// Padding Vertical 36
+  const {{class_name.pascalCase()}}Padding.vertical36() : super.symmetric(vertical: 36);
+
+  /// Padding Vertical 40
+  const {{class_name.pascalCase()}}Padding.vertical40() : super.symmetric(vertical: 40);
+
+  /// Padding Only Start
+  /// Padding Only Start 4
+  const {{class_name.pascalCase()}}Padding.onlyStart4() : super.only(start: 4);
+
+  /// Padding Only Start 8
+  const {{class_name.pascalCase()}}Padding.onlyStart8() : super.only(start: 8);
+
+  /// Padding Only Start 12
+  const {{class_name.pascalCase()}}Padding.onlyStart12() : super.only(start: 12);
+
+  /// Padding Only Start 16
+  const {{class_name.pascalCase()}}Padding.onlyStart16() : super.only(start: 16);
+
+  /// Padding Only Start 20
+  const {{class_name.pascalCase()}}Padding.onlyStart20() : super.only(start: 20);
+
+  /// Padding Only Start 24
+  const {{class_name.pascalCase()}}Padding.onlyStart24() : super.only(start: 24);
+
+  /// Padding Only Start 28
+  const {{class_name.pascalCase()}}Padding.onlyStart28() : super.only(start: 28);
+
+  /// Padding Only Start 32
+  const {{class_name.pascalCase()}}Padding.onlyStart32() : super.only(start: 32);
+
+  /// Padding Only Start 36
+  const {{class_name.pascalCase()}}Padding.onlyStart36() : super.only(start: 36);
+
+  /// Padding Only Start 40
+  const {{class_name.pascalCase()}}Padding.onlyStart40() : super.only(start: 40);
+
+  /// Padding Only End
+  /// Padding Only End 4
+  const {{class_name.pascalCase()}}Padding.onlyEnd4() : super.only(end: 4);
+
+  /// Padding Only End 8
+  const {{class_name.pascalCase()}}Padding.onlyEnd8() : super.only(end: 8);
+
+  /// Padding Only End 12
+  const {{class_name.pascalCase()}}Padding.onlyEnd12() : super.only(end: 12);
+
+  /// Padding Only End 16
+  const {{class_name.pascalCase()}}Padding.onlyEnd16() : super.only(end: 16);
+
+  /// Padding Only End 20
+  const {{class_name.pascalCase()}}Padding.onlyEnd20() : super.only(end: 20);
+
+  /// Padding Only End 24
+  const {{class_name.pascalCase()}}Padding.onlyEnd24() : super.only(end: 24);
+
+  /// Padding Only End 28
+  const {{class_name.pascalCase()}}Padding.onlyEnd28() : super.only(end: 28);
+
+  /// Padding Only End 32
+  const {{class_name.pascalCase()}}Padding.onlyEnd32() : super.only(end: 32);
+
+  /// Padding Only End 36
+  const {{class_name.pascalCase()}}Padding.onlyEnd36() : super.only(end: 36);
+
+  /// Padding Only End 40
+  const {{class_name.pascalCase()}}Padding.onlyEnd40() : super.only(end: 40);
+
+  /// Padding Only Top
+  /// Padding Only Top 4
+  const {{class_name.pascalCase()}}Padding.onlyTop4() : super.only(top: 4);
+
+  /// Padding Only Top 8
+  const {{class_name.pascalCase()}}Padding.onlyTop8() : super.only(top: 8);
+
+  /// Padding Only Top 12
+  const {{class_name.pascalCase()}}Padding.onlyTop12() : super.only(top: 12);
+
+  /// Padding Only Top 16
+  const {{class_name.pascalCase()}}Padding.onlyTop16() : super.only(top: 16);
+
+  /// Padding Only Top 20
+  const {{class_name.pascalCase()}}Padding.onlyTop20() : super.only(top: 20);
+
+  /// Padding Only Top 24
+  const {{class_name.pascalCase()}}Padding.onlyTop24() : super.only(top: 24);
+
+  /// Padding Only Top 28
+  const {{class_name.pascalCase()}}Padding.onlyTop28() : super.only(top: 28);
+
+  /// Padding Only Top 32
+  const {{class_name.pascalCase()}}Padding.onlyTop32() : super.only(top: 32);
+
+  /// Padding Only Top 36
+  const {{class_name.pascalCase()}}Padding.onlyTop36() : super.only(top: 36);
+
+  /// Padding Only Top 40
+  const {{class_name.pascalCase()}}Padding.onlyTop40() : super.only(top: 40);
+
+  /// Padding Only Bottom
+  /// Padding Only Bottom 4
+  const {{class_name.pascalCase()}}Padding.onlyBottom4() : super.only(bottom: 4);
+
+  /// Padding Only Bottom 8
+  const {{class_name.pascalCase()}}Padding.onlyBottom8() : super.only(bottom: 8);
+
+  /// Padding Only Bottom 12
+  const {{class_name.pascalCase()}}Padding.onlyBottom12() : super.only(bottom: 12);
+
+  /// Padding Only Bottom 16
+  const {{class_name.pascalCase()}}Padding.onlyBottom16() : super.only(bottom: 16);
+
+  /// Padding Only Bottom 20
+  const {{class_name.pascalCase()}}Padding.onlyBottom20() : super.only(bottom: 20);
+
+  /// Padding Only Bottom 24
+  const {{class_name.pascalCase()}}Padding.onlyBottom24() : super.only(bottom: 24);
+
+  /// Padding Only Bottom 28
+  const {{class_name.pascalCase()}}Padding.onlyBottom28() : super.only(bottom: 28);
+
+  /// Padding Only Bottom 32
+  const {{class_name.pascalCase()}}Padding.onlyBottom32() : super.only(bottom: 32);
+
+  /// Padding Only Bottom 36
+  const {{class_name.pascalCase()}}Padding.onlyBottom36() : super.only(bottom: 36);
+
+  /// Padding Only Bottom 40
+  const {{class_name.pascalCase()}}Padding.onlyBottom40() : super.only(bottom: 40);
+}