+++ bricks/ui_kit/__brick__/lib/layout/radius/{{class_name.snakeCase()}}_border_radius.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:{{class_name.snakeCase()}}/layout/radius/{{class_name.snakeCase()}}_radius.dart';
+
+class {{class_name.pascalCase()}}BorderRadius extends BorderRadius {
+  /// Circular Zero
+  const {{class_name.pascalCase()}}BorderRadius.zero() : super.all(const {{class_name.pascalCase()}}Radius.zero());
+
+  /// Circular 4
+  const {{class_name.pascalCase()}}BorderRadius.v4() : super.all(const {{class_name.pascalCase()}}Radius.v4());
+
+  /// Circular 8
+  const {{class_name.pascalCase()}}BorderRadius.v8() : super.all(const {{class_name.pascalCase()}}Radius.v8());
+
+  /// Circular 12
+  const {{class_name.pascalCase()}}BorderRadius.v12() : super.all(const {{class_name.pascalCase()}}Radius.v12());
+
+  /// Circular 16
+  const {{class_name.pascalCase()}}BorderRadius.v16() : super.all(const {{class_name.pascalCase()}}Radius.v16());
+
+  /// Circular 20
+  const {{class_name.pascalCase()}}BorderRadius.v20() : super.all(const {{class_name.pascalCase()}}Radius.v20());
+
+  /// Circular 24
+  const {{class_name.pascalCase()}}BorderRadius.v24() : super.all(const {{class_name.pascalCase()}}Radius.v24());
+
+  /// Circular 28
+  const {{class_name.pascalCase()}}BorderRadius.v28() : super.all(const {{class_name.pascalCase()}}Radius.v28());
+
+  /// Circular 32
+  const {{class_name.pascalCase()}}BorderRadius.v32() : super.all(const {{class_name.pascalCase()}}Radius.v32());
+
+  /// Circular 36
+  const {{class_name.pascalCase()}}BorderRadius.v36() : super.all(const {{class_name.pascalCase()}}Radius.v36());
+
+  /// Circular 40
+  const {{class_name.pascalCase()}}BorderRadius.v40() : super.all(const {{class_name.pascalCase()}}Radius.v40());
+}