+++ bricks/ui_kit/__brick__/lib/layout/radius/{{class_name.snakeCase()}}_radius.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class {{class_name.pascalCase()}}Radius extends Radius {
+  /// Circular Zero
+  const {{class_name.pascalCase()}}Radius.zero() : super.circular(0);
+
+  /// Circular 4
+  const {{class_name.pascalCase()}}Radius.v4() : super.circular(4);
+
+  /// 8
+  const {{class_name.pascalCase()}}Radius.v8() : super.circular(8);
+
+  /// Circular 12
+  const {{class_name.pascalCase()}}Radius.v12() : super.circular(12);
+
+  /// Circular 16
+  const {{class_name.pascalCase()}}Radius.v16() : super.circular(16);
+
+  /// Circular 20
+  const {{class_name.pascalCase()}}Radius.v20() : super.circular(20);
+
+  /// Circular 24
+  const {{class_name.pascalCase()}}Radius.v24() : super.circular(24);
+
+  /// Circular 28
+  const {{class_name.pascalCase()}}Radius.v28() : super.circular(28);
+
+  /// 32
+  const {{class_name.pascalCase()}}Radius.v32() : super.circular(32);
+
+  /// Circular 36
+  const {{class_name.pascalCase()}}Radius.v36() : super.circular(36);
+
+  /// 40
+  const {{class_name.pascalCase()}}Radius.v40() : super.circular(40);
+}