+++ bricks/ui_kit/__brick__/lib/theme/color/color.dart
@@ -1,3 +1,3 @@
-export 'ui_kit_colors.dart';
-export 'ui_kit_dark_colors.dart';
-export 'ui_kit_light_colors.dart';
+export '{{class_name.snakeCase()}}_colors.dart';
+export '{{class_name.snakeCase()}}_dark_colors.dart';
+export '{{class_name.snakeCase()}}_light_colors.dart';