+++ bricks/ui_kit/__brick__/lib/theme/extensions/extensions.dart
@@ -1,1 +1,1 @@
-export 'ui_kit_theme_extensions.dart';
+export '{{class_name.snakeCase()}}_theme_extensions.dart';