+++ bricks/ui_kit/__brick__/lib/ui_kit.dart
@@ -1,7 +1,0 @@
-///
-library ui_kit;
-
-export 'constants/ui_kit_constants.dart';
-export 'layout/layout.dart';
-export 'theme/theme.dart';
-export 'widgets/widgets.dart';