+++ bricks/ui_kit/__brick__/lib/{{class_name.snakeCase()}}.dart
@@ -1,0 +1,7 @@
+///
+library {{class_name.snakeCase()}};
+
+export 'constants/{{class_name.snakeCase()}}_constants.dart';
+export 'layout/layout.dart';
+export 'theme/theme.dart';
+export 'widgets/widgets.dart';