+++ examples/ui_kit/lib/layout/gap/ui_kit_v_gap.dart
@@ -1,37 +1,37 @@
 import 'package:flutter/material.dart';
 
-class UIKitVGap extends SizedBox {
+class UiKitVGap extends SizedBox {
   /// Vertical Gap Zero
-  const UIKitVGap({super.key}) : super(height: 0);
-  const UIKitVGap.zero({super.key}) : super(height: 0);
+  const UiKitVGap({super.key}) : super(height: 0);
+  const UiKitVGap.zero({super.key}) : super(height: 0);
 
   /// Vertical Gap 4
-  const UIKitVGap.v4({super.key}) : super(height: 4);
+  const UiKitVGap.v4({super.key}) : super(height: 4);
 
   /// Vertical Gap 8
-  const UIKitVGap.v8({super.key}) : super(height: 8);
+  const UiKitVGap.v8({super.key}) : super(height: 8);
 
   /// Vertical Gap 12
-  const UIKitVGap.v12({super.key}) : super(height: 12);
+  const UiKitVGap.v12({super.key}) : super(height: 12);
 
   /// Vertical Gap 16
-  const UIKitVGap.v16({super.key}) : super(height: 16);
+  const UiKitVGap.v16({super.key}) : super(height: 16);
 
   /// Vertical Gap 20
-  const UIKitVGap.v20({super.key}) : super(height: 20);
+  const UiKitVGap.v20({super.key}) : super(height: 20);
 
   /// Vertical Gap 24
-  const UIKitVGap.v24({super.key}) : super(height: 24);
+  const UiKitVGap.v24({super.key}) : super(height: 24);
 
   /// Vertical Gap 28
-  const UIKitVGap.v28({super.key}) : super(height: 28);
+  const UiKitVGap.v28({super.key}) : super(height: 28);
 
   /// Vertical Gap 32
-  const UIKitVGap.v32({super.key}) : super(height: 32);
+  const UiKitVGap.v32({super.key}) : super(height: 32);
 
   /// Vertical Gap 36
-  const UIKitVGap.v36({super.key}) : super(height: 36);
+  const UiKitVGap.v36({super.key}) : super(height: 36);
 
   /// Vertical Gap 40
-  const UIKitVGap.v40({super.key}) : super(height: 40);
+  const UiKitVGap.v40({super.key}) : super(height: 40);
 }