+++ examples/ui_kit/lib/layout/radius/ui_kit_radius.dart
@@ -1,36 +1,36 @@
 import 'package:flutter/material.dart';
 
-class UIKitRadius extends Radius {
+class UiKitRadius extends Radius {
   /// Circular Zero
-  const UIKitRadius.zero() : super.circular(0);
+  const UiKitRadius.zero() : super.circular(0);
 
   /// Circular 4
-  const UIKitRadius.v4() : super.circular(4);
+  const UiKitRadius.v4() : super.circular(4);
 
   /// 8
-  const UIKitRadius.v8() : super.circular(8);
+  const UiKitRadius.v8() : super.circular(8);
 
   /// Circular 12
-  const UIKitRadius.v12() : super.circular(12);
+  const UiKitRadius.v12() : super.circular(12);
 
   /// Circular 16
-  const UIKitRadius.v16() : super.circular(16);
+  const UiKitRadius.v16() : super.circular(16);
 
   /// Circular 20
-  const UIKitRadius.v20() : super.circular(20);
+  const UiKitRadius.v20() : super.circular(20);
 
   /// Circular 24
-  const UIKitRadius.v24() : super.circular(24);
+  const UiKitRadius.v24() : super.circular(24);
 
   /// Circular 28
-  const UIKitRadius.v28() : super.circular(28);
+  const UiKitRadius.v28() : super.circular(28);
 
   /// 32
-  const UIKitRadius.v32() : super.circular(32);
+  const UiKitRadius.v32() : super.circular(32);
 
   /// Circular 36
-  const UIKitRadius.v36() : super.circular(36);
+  const UiKitRadius.v36() : super.circular(36);
 
   /// 40
-  const UIKitRadius.v40() : super.circular(40);
+  const UiKitRadius.v40() : super.circular(40);
 }