+++ examples/ui_kit/CHANGELOG.md
@@ -1,3 +1,2 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+- ui_kit package created.